push more primitives into the layers (#354)
* embeds -> 8mi, rm keras/dy exporters, cleanup

* fix broken stuff in pytorch tagger

* make the eager examples run properly again

* push down CM

* Move more stuff around
