+++ python/baseline/__init__.py
@@ -1,5 +1,4 @@
 try:
-    from baseline.mime_type import *
     from baseline.utils import *
     from baseline.vectorizers import *
     from baseline.w2v import *
@@ -12,9 +11,8 @@ try:
     from baseline.embeddings import *
     from baseline.services import *
     from baseline.train import *
-
     logger = get_console_logger('baseline', env_key='BASELINE_LOG_LEVEL')
     report_logger = get_console_logger('baseline.reporting', env_key='BASELINE_LOG_LEVEL')
-except ImportError:
-    pass
+except ImportError as e:
+    print(e)
 from baseline.version import __version__
