+++ python/baseline/pytorch/classify/train.py
@@ -4,15 +4,17 @@ import torch
 import torch.autograd
 import os
 
-from baseline.utils import verbose_output
-from baseline.confusion import ConfusionMatrix
-from baseline.progress import create_progress_bar
-from baseline.utils import listify, get_model_file, get_metric_cmp
-from baseline.pytorch.optz import OptimizerManager
+from eight_mile.confusion import ConfusionMatrix
+from eight_mile.progress import create_progress_bar
+from eight_mile.utils import listify
+from eight_mile.pytorch.optz import OptimizerManager
+
+from baseline.utils import verbose_output, get_model_file, get_metric_cmp
 from baseline.train import EpochReportingTrainer, create_trainer, register_trainer, register_training_func
 from baseline.model import create_model_for
 
 logger = logging.getLogger('baseline')
+
 
 def _add_to_cm(cm, y, pred):
     _, best = pred.max(1)
