+++ python/baseline/pytorch/lm/train.py
@@ -1,11 +1,13 @@
 import time
 import logging
 from baseline.pytorch.torchy import *
-from baseline.utils import listify, revlut, get_model_file, get_metric_cmp
+from eight_mile.utils import listify, revlut
+from eight_mile.pytorch.optz import OptimizerManager
+from baseline.utils import get_model_file, get_metric_cmp
 from baseline.train import Trainer, create_trainer, register_trainer, register_training_func
-from baseline.pytorch.optz import OptimizerManager
 from baseline.model import create_model_for
 logger = logging.getLogger('baseline')
+
 
 @register_trainer(task='lm', name='default')
 class LanguageModelTrainerPyTorch(Trainer):
