--- python/mead/dy/README.md
@@ -1,1 +1,0 @@
-# Exporting Dynet models is not supported.
