--- python/mead/keras/README.md
@@ -1,1 +1,0 @@
-# Exporting Keras models is not supported.
